@@ -4,7 +4,14 @@
   "description": "Roost visualization and annotation web interface",
   "main": "vis.js",
   "scripts": {
-    "build": "webpack --config webpack.config.js"
+    "build": "npm-run-all -s build:rollup build:rollup:dist",
+    "build:rollup": "rollup -c",
+    "build:rollup:dist": "rollup -c --environment BUILD:production",
+    "watch": "npm-run-all -s build -p watch:rollup watch:rollup:dist",
+    "watch:rollup": "rollup -w -c",
+    "watch:rollup:dist": "rollup -w -c --environment BUILD:production",
+    "serve": "http-server -p 8888 --cors",
+    "watch-serve": "npm-run-all -p serve watch"
   },
   "repository": {
     "type": "git",
@@ -18,24 +25,38 @@
   },
   "homepage": "https://github.com/darkecology/roostui#readme",
   "dependencies": {
+    "@babel/preset-es2015": "^7.0.0-beta.53",
     "d3": "^6.6.0",
     "d3-array": "^2.12.0",
     "d3-dsv": "^2.0.0",
+    "debug": "^4.3.1",
     "http-browserify": "^1.7.0",
     "jquery": "^3.6.0",
-    "plotly": "^1.0.6",
-    "plotly.js-dist-min": "^1.58.4",
+    "npm-run-all": "^4.1.5",
+    "rollup": "^2.42.0",
+    "rollup-plugin-babel": "^4.4.0",
     "sprintf": "^0.1.5",
-    "webpack-cli": "^4.5.0"
+    "webpack": "^5.27.1"
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
+    "@babel/plugin-external-helpers": "^7.12.13",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-transform-runtime": "^7.13.10",
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-react": "^7.12.13",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-buble": "^0.21.3",
+    "@rollup/plugin-commonjs": "^17.1.0",
+    "@rollup/plugin-node-resolve": "^11.2.0",
     "babel-loader": "^8.2.2",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.25.0"
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-sourcemaps": "^0.6.3",
+    "rollup-plugin-uglify": "^6.0.4",
+    "url-loader": "^4.1.1"
   },
   "browser": {
     "http": false,
